Add files via upload
<div style="text-align: justify; text-justify: inter-word;"> </div> 

Justified text is spaced so the left and right sides of the text block both have a clean edge. The usual alternative to justified text is left-aligned text , which has a straight left edge and an uneven right edge. Compared to left-aligned text, justification gives text a cleaner, more formal look.